+++ bzt/resources/chrome_launcher.sh
@@ -1,6 +1,3 @@
 #!/bin/bash
 
-WRAPPER=$(readlink -f "$0")
-HERE=$(dirname "$WRAPPER")
-
-"$HERE/_google-chrome" --no-sandbox "$@"
+exec "/opt/google/chrome/google-chrome" --no-sandbox "$@"