+++ site/dat/docs/Changelog.md
@@ -3,6 +3,7 @@
 ## 1.7.4 (next)
  - fix Locust crash when used with 'requests'-style scenario and cloud provisioning
  - fix JSR223 block handling when using cloud/remote provisioning
+ - do not generate Selenium scripts for cloud tests
 
 ## 1.7.3 <sup>1 nov 2016</sup>
  - add TestNG-based runner for Selenium tests